Split tool auth context utils into one export per file
Review feedback: move buildRequiredToolAuthContext to its own util file and
the shared dependencies type to a type file.
