+++ packages/twenty-server/src/engine/core-modules/tool-provider/types/tool-auth-context-dependencies.type.ts
@@ -1,0 +1,9 @@
+import { type Repository } from 'typeorm';
+
+import { type UserEntity } from 'src/engine/core-modules/user/user.entity';
+import { type WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
+
+export type ToolAuthContextDependencies = {
+  userRepository: Pick<Repository<UserEntity>, 'findOne'>;
+  workspaceCacheService: Pick<WorkspaceCacheService, 'getOrRecompute'>;
+};
