+++ packages/twenty-server/src/engine/core-modules/tool-provider/utils/__tests__/with-resolved-tool-auth-context.util.spec.ts
@@ -1,14 +1,10 @@
-import { AuthException } from 'src/engine/core-modules/auth/auth.exception';
 import {
   getWorkspaceAuthContext,
   workspaceAuthContextStorage,
 } from 'src/engine/core-modules/auth/storage/workspace-auth-context.storage';
 import { type WorkspaceAuthContext } from 'src/engine/core-modules/auth/types/workspace-auth-context.type';
 import { type ToolProviderContext } from 'src/engine/core-modules/tool-provider/interfaces/tool-provider-context.type';
-import {
-  buildRequiredToolAuthContext,
-  withResolvedToolAuthContext,
-} from 'src/engine/core-modules/tool-provider/utils/with-resolved-tool-auth-context.util';
+import { withResolvedToolAuthContext } from 'src/engine/core-modules/tool-provider/utils/with-resolved-tool-auth-context.util';
 
 describe('withResolvedToolAuthContext', () => {
   const workspaceId = '20202020-0000-4000-8000-000000000001';
@@ -164,60 +160,5 @@ describe('withResolvedToolAuthContext', () => {
     );
 
     expect(result).toEqual({ success: true });
-  });
-});
-
-describe('buildRequiredToolAuthContext', () => {
-  const workspaceId = '20202020-0000-4000-8000-000000000001';
-  const userId = '20202020-0000-4000-8000-000000000002';
-  const userWorkspaceId = '20202020-0000-4000-8000-000000000003';
-
-  const buildContext = (
-    overrides?: Partial<ToolProviderContext>,
-  ): ToolProviderContext => ({
-    workspaceId,
-    roleId: 'role-1',
-    rolePermissionConfig: { unionOf: ['role-1'] },
-    ...overrides,
-  });
-
-  it('should throw without a resolvable identity', async () => {
-    await expect(
-      buildRequiredToolAuthContext({
-        context: buildContext(),
-        userRepository: { findOne: jest.fn() },
-        workspaceCacheService: { getOrRecompute: jest.fn() },
-      }),
-    ).rejects.toThrow(AuthException);
-  });
-
-  it('should throw when the user cannot be found', async () => {
-    await expect(
-      buildRequiredToolAuthContext({
-        context: buildContext({ userId, userWorkspaceId }),
-        userRepository: { findOne: jest.fn().mockResolvedValue(null) },
-        workspaceCacheService: { getOrRecompute: jest.fn() },
-      }),
-    ).rejects.toThrow('User not found');
-  });
-
-  it('should throw when the user has no workspace member', async () => {
-    await expect(
-      buildRequiredToolAuthContext({
-        context: buildContext({ userId, userWorkspaceId }),
-        userRepository: {
-          findOne: jest.fn().mockResolvedValue({
-            id: userId,
-            createdAt: new Date('2024-01-01T00:00:00Z'),
-            updatedAt: new Date('2024-01-01T00:00:00Z'),
-          }),
-        },
-        workspaceCacheService: {
-          getOrRecompute: jest.fn().mockResolvedValue({
-            flatWorkspaceMemberMaps: { idByUserId: {}, byId: {} },
-          }),
-        },
-      }),
-    ).rejects.toThrow('Workspace member not found');
   });
 });
